Update script.js
When we made the repo private, it inactivated our main branch protective rules. So, I made the decision to delete the API key form the JS file and set the visibility back to the published site. This seems to be the only way to keep the branch protection in place.